+++ api/invite_test.go
@@ -673,7 +673,7 @@ func TestInviteResponds(t *testing.T) {
 		{
 			desc:     "valid request to accept an team invite",
 			method:   http.MethodPut,
-			url:      fmt.Sprintf("/accept/team/invite/%s/%s", "123.456.789", "123456"),
+			url:      "/accept/team/invite",
 			token:    testing_token_uid1,
 			respCode: http.StatusOK,
 			body: testJSONObject{
@@ -683,7 +683,7 @@ func TestInviteResponds(t *testing.T) {
 		{
 			desc:     "valid request to accept an team invite for a patient",
 			method:   http.MethodPut,
-			url:      fmt.Sprintf("/accept/team/invite/%s/%s", "123.456.789", "123456"),
+			url:      "/accept/team/invite",
 			token:    testing_token_uid1,
 			respCode: http.StatusOK,
 			body: testJSONObject{
@@ -692,19 +692,19 @@ func TestInviteResponds(t *testing.T) {
 			},
 		},
 		{
-			desc:     "not authorized request to accept an team invite",
+			desc:     "not authorized request to accept a team invite",
 			method:   http.MethodPut,
-			url:      fmt.Sprintf("/accept/team/invite/%s/%s", "not.authorized", "123456"),
+			url:      "/accept/team/invite",
 			token:    testing_token_uid1,
-			respCode: http.StatusUnauthorized,
+			respCode: http.StatusForbidden,
 			body: testJSONObject{
-				"key": "medicalteam.invite.member",
+				"key": "medicalteam.invite.wrong.member",
 			},
 		},
 		{
 			desc:     "invitation does not exist",
 			method:   http.MethodPut,
-			url:      fmt.Sprintf("/accept/team/invite/%s/%s", "123.456.789", "123456"),
+			url:      "/accept/team/invite",
 			token:    testing_token_uid1,
 			respCode: http.StatusForbidden,
 			body: testJSONObject{
@@ -714,12 +714,80 @@ func TestInviteResponds(t *testing.T) {
 		{
 			desc:     "invalid invitation",
 			method:   http.MethodPut,
-			url:      fmt.Sprintf("/accept/team/invite/%s/%s", "123.456.789", "123456"),
+			url:      "/accept/team/invite",
 			token:    testing_token_uid1,
 			respCode: http.StatusNotFound,
 			body: testJSONObject{
 				"key": "key.does.not.exist",
 			},
+		},
+		{
+			desc:     "Any invite no key",
+			method:   http.MethodPut,
+			url:      "/accept/team/invite",
+			token:    testing_token_uid1,
+			respCode: http.StatusBadRequest,
+		},
+		{
+			desc:   "Any invite invalid key",
+			method: http.MethodPut,
+			url:    "/accept/team/invite",
+			token:  testing_token_uid1,
+			body: testJSONObject{
+				"key": "any.invite.invalid.key",
+			},
+			respCode: http.StatusNotFound,
+		},
+		{
+			desc:   "Any invite already completed",
+			method: http.MethodPut,
+			url:    "/accept/team/invite",
+			token:  testing_token_uid1,
+			body: testJSONObject{
+				"key": "any.invite.completed.key",
+			},
+			respCode: http.StatusForbidden,
+		},
+		{
+			desc:   "Error getting invite",
+			doBad:  true,
+			method: http.MethodPut,
+			url:    "/accept/team/invite",
+			token:  testing_token_uid1,
+			body: testJSONObject{
+				"key": "foo",
+			},
+			respCode: http.StatusInternalServerError,
+		},
+		{
+			desc:   "Any invite not a valid type",
+			method: http.MethodPut,
+			url:    "/accept/team/invite",
+			token:  testing_token_uid1,
+			body: testJSONObject{
+				"key": "invite.wrong.type",
+			},
+			respCode: http.StatusForbidden,
+		},
+		{
+			desc:   "Any valid invite do admin",
+			method: http.MethodPut,
+			url:    "/accept/team/invite",
+			token:  testing_token_uid1,
+			body: testJSONObject{
+				"key": "any.invite.pending.do.admin",
+			},
+			respCode: http.StatusOK,
+		},
+		{
+			desc:   "Any valid invite remove",
+			method: http.MethodPut,
+			url:    "/accept/team/invite",
+			token:  testing_token_uid1,
+			body: testJSONObject{
+				"key": "any.invite.pending.remove",
+			},
+			respCode: http.StatusOK,
 		},
 	}
 
@@ -767,6 +835,19 @@ func TestInviteResponds(t *testing.T) {
 			hydrophone = InitApi(
 				FAKE_CONFIG,
 				mockStoreEmpty,
+				mockNotifier,
+				mockShoreline,
+				mockPerms,
+				mockSeagull,
+				mockPortal,
+				mockTemplates,
+			)
+		}
+		// testing when returning errors
+		if inviteTest.doBad {
+			hydrophone = InitApi(
+				FAKE_CONFIG,
+				mockStoreFails,
 				mockNotifier,
 				mockShoreline,
 				mockPerms,