+++ models/confirmation.go
@@ -183,16 +183,6 @@ func (c *Confirmation) ValidateUserID(expectedUserID string, validationErrors *[
 	return c
 }
 
-func (c *Confirmation) ValidateTeamID(expectedTeamID string, validationErrors *[]error) *Confirmation {
-	if expectedTeamID != c.TeamID {
-		*validationErrors = append(
-			*validationErrors,
-			fmt.Errorf("Confirmation expected TeamId of `%s` but had `%s`", expectedTeamID, c.TeamID),
-		)
-	}
-	return c
-}
-
 func (c *Confirmation) ValidateStatus(expectedStatus Status, validationErrors *[]error) *Confirmation {
 	if expectedStatus != c.Status {
 		*validationErrors = append(
@@ -203,13 +193,18 @@ func (c *Confirmation) ValidateStatus(expectedStatus Status, validationErrors *[
 	return c
 }
 
-func (c *Confirmation) ValidateType(expectedType Type, validationErrors *[]error) *Confirmation {
-	if expectedType != c.Type {
-		*validationErrors = append(
-			*validationErrors,
-			fmt.Errorf("Confirmation expected Type `%s` but had `%s`", expectedType, c.Type),
-		)
+func (c *Confirmation) ValidateType(expectedTypes []Type, validationErrors *[]error) *Confirmation {
+	types := ""
+	for _, expectedType := range expectedTypes {
+		if expectedType == c.Type {
+			return c
+		}
+		types = fmt.Sprintf("%s, %s", types, expectedType)
 	}
+	*validationErrors = append(
+		*validationErrors,
+		fmt.Errorf("Confirmation expected Type(s) `%s` but had `%s`", types, c.Type),
+	)
 	return c
 }